+++ R/data.R
@@ -6,11 +6,11 @@
 #' 
 #'\tabular{ll}{
 #'  \code{employment}  \tab (EM) \cr
-#'  \code{FE}          \tab further education (FE)\cr
+#'  \code{FE}          \tab further education (FE) \cr
 #'  \code{HE}          \tab higher education (HE) \cr
 #'  \code{joblessness} \tab (JL) \cr
 #'  \code{school}      \tab (SC) \cr
-#'  \code{training}    \tab (TR)\cr
+#'  \code{training}    \tab (TR) \cr
 #'}
 
 #'The data set contains also ids (\code{id}) and sample weights (\code{weights}) as well as the following binary covariates:\cr
@@ -55,18 +55,18 @@
 #' 
 #' The covariates are: \cr
 #' \tabular{ll}{
-#' \code{sex} \tab \cr
-#' \code{birthyr }  \tab (birth year)\cr
-#' \code{nat_1_02}  \tab (first nationality) \cr
-#' \code{plingu02}  \tab (language of questionnaire) \cr
-#' \code{p02r01 }   \tab (religion)\cr
-#' \code{p02r04 }   \tab (religious participation) \cr
-#' \code{cspfaj }   \tab (father's social status) \cr
-#' \code{cspmoj }   \tab (mother's social status) \cr
+#' \code{sex}      \tab \cr
+#' \code{birthyr}  \tab (birth year) \cr
+#' \code{nat_1_02} \tab (first nationality) \cr
+#' \code{plingu02} \tab (language of questionnaire) \cr
+#' \code{p02r01}   \tab (religion) \cr
+#' \code{p02r04}   \tab (religious participation) \cr
+#' \code{cspfaj}   \tab (father's social status) \cr
+#' \code{cspmoj}   \tab (mother's social status) \cr
 #' }
 #' Two additional weights variables are inserted for illustrative purpose ONLY (since \code{biofam} is a subsample of the original data, these weights are not adapted to the actual data): \cr
 #' \tabular{ll}{
-#' \code{wp00tbgp} \tab (weights inflating to the Swiss population)\cr
+#' \code{wp00tbgp} \tab (weights inflating to the Swiss population) \cr
 #' \code{wp00tbgs} \tab (weights respecting sample size) \cr
 #'}
 #' @references Mueller, N. S., Studer, M. and Ritschard, G. (2007). Classification de parcours de vie a l'aide de l'optimal matching. In \emph{XIVe Rencontre de la Societe francophone de classification (SFC 2007), Paris, 5-7 septembre 2007}, pp. 157-160.